+++ dice/src/test/kotlin/com/embabel/dice/proposition/extraction/ExtractionRunValueTypesTest.kt
@@ -233,6 +233,7 @@ class ExtractionRunValueTypesTest {
             ExtractionRunFingerprint::class.java,
             ExtractionRunNotFoundException::class.java,
             ExtractionRunConflictException::class.java,
+            com.embabel.dice.common.ExtractionRunTransitioned::class.java,
         ).forEach { type ->
             assertThat(isMarkedExperimental(type))
                 .describedAs("%s is marked experimental", type.simpleName)